% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_package.R
\docType{data}
\name{ex_dta}
\alias{ex_dta}
\title{Example dataset

Example dataset of a single arm study.}
\format{
A data frame with the following variables:
\itemize{
  \item{Enroll_Time}{Days from the time the first patient was enrolled}
  \item{Y}{Binary outcome}
}
}
\usage{
data(ex_dta)
}
\description{
Example dataset

Example dataset of a single arm study.
}
\keyword{datasets}
