% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_package.R
\name{parameters}
\alias{parameters}
\title{Parameters}
\arguments{
\item{epsilon}{mixing proportion from the posterior distribution}

\item{nsmps}{number of samples to draw}

\item{mix_ab}{shape parameters for Beta}
}
\description{
Parameters
}
