% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_combine}
\alias{pd_ss_combine}
\title{Combine simulation study results}
\usage{
pd_ss_combine(prefix, cmb_reps = 1:50, f_rst = "rst.Rdata")
}
\description{
Combine simulation study results
}
