% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_get_para}
\alias{pd_ss_get_para}
\title{Conduct logistic regression with a parametric model}
\usage{
pd_ss_get_para(dta_pts, nx = 0)
}
\arguments{
\item{nx}{number of covariates to put in the model}
}
\description{

}
