% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_get_reference}
\alias{pd_ss_get_reference}
\title{Draw inference for Epsilon = 0 and 1}
\usage{
pd_ss_get_reference(dta_sum, labels = c("Eps = 0", "Eps = 1"))
}
\arguments{
\item{dta_sum}{summary of the simulated patients, a return from
\code{pd_ss_patients}}
}
\description{
Draw inference for Epsilon = 0 and 1
}
