% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_plt_bayes}
\alias{pd_ss_plt_bayes}
\title{Generate plot for posterior distributions}
\usage{
pd_ss_plt_bayes(rst_bayesian, rst_summary, n_interval = 4)
}
\description{
Generate plot for posterior distributions
}
