% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_plt_truth}
\alias{pd_ss_plt_truth}
\title{Plot observed response rates}
\usage{
pd_ss_plt_truth(
  dtaSum,
  fname = "plt_interval.pdf",
  ylim = c(0.25, 0.36),
  xlim = c(0, 3)
)
}
\description{
Plot observed response rates
}
