% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_summarize}
\alias{pd_ss_summarize}
\title{Summarize simulation study results}
\usage{
pd_ss_summarize(
  rst_truth,
  rst_theta,
  rst_filter,
  rst_par,
  rst_par0,
  rst_trt,
  n_current,
  cur_set = NULL
)
}
\description{
Summarize simulation study results
}
