% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{pd_ss_truth}
\alias{pd_ss_truth}
\title{Simulation Truth}
\usage{
pd_ss_truth(simu_setting, ntruth = 1e+06, ...)
}
\arguments{
\item{simu_setting}{A list of simulation parameters for the following
functions \code{pdSimuPts}, \code{pdGetInt}}

\item{ntruth}{number of patients to simulate for summarizing the truth}
}
\description{
Simulation Truth
}
\examples{
    simu_setting <- list(n_pat      = 100,
                         mu_0       = -1.5,
                         mu_0_trt   = -0.4,
                         fmla_t     = as.formula("~ -1 + time"),
                         coeff_t    = 0,
                         muCov      = rep(0,   5)
                         sdCov      = rep(0.5, 5)
                         corCov     = 0.1,
                         coeff_x    = rep(0.5, 5)
                         t_end      = 3,
                         t_start    = 0,
                         t_hist     = 2,
                         intervals  = 2,
                         seed       = 1000)

   simu_patients(simu_setting)

}
