% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_filter.R
\name{pdBinPost}
\alias{pdBinPost}
\title{Posterior Beta}
\usage{
pdBinPost(y, nsmps = 5000, mix_ab = c(0.25, 0.25), ...)
}
\arguments{
\item{nsmps}{number of samples to draw}

\item{mix_ab}{shape parameters for Beta}
}
\description{
Posterior Beta
}
