% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_parametric.R
\name{pdESS}
\alias{pdESS}
\title{Effective Sample Size}
\usage{
pdESS(dist_post, dist_ref, n_ref)
}
\arguments{
\item{dist_post}{posterior distribution based on informative prior}

\item{dist_ref}{posterior distribution based on non-informative prior}

\item{n_ref}{sample size in the current study}
}
\value{
Effective sample size
}
\description{
Effective Sample Size
}
