% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_filter.R
\name{pdFilter}
\alias{pdFilter}
\title{Particle filter all}
\usage{
pdFilter(
  vec_y,
  vec_interval,
  nsmps = 5000,
  epsilons = 1,
  f_tran = pdTran,
  f_post = pdBinPost,
  f_pi0 = pdBinPi0,
  f_filter = pdFilterSingle,
  update_eps = FALSE,
  ...
)
}
\arguments{
\item{f_tran}{transition function}

\item{f_post}{posterior sampling when the prior is pi_0, usually for the
first interval}

\item{f_pi0}{pi_0 function}

\item{f_filter}{particle filter function}

\item{update_eps}{whether keep epsilon as constant (design parameter)}

\item{...}{parameters for the transition function}

\item{vecy}{vector of binary outcomes}

\item{mix_ab}{beta shape parameters (a, b) for a non-informative prior}

\item{epsillon}{mixture proportion from the prior samples.}
}
\description{
Particle filter all
}
