% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simu.R
\name{pdGetInt}
\alias{pdGetInt}
\title{Create time interval}
\usage{
pdGetInt(
  vec_t,
  n_interval = 4,
  t_start = 0,
  t_hist = 2,
  t_end = 3,
  remap = FALSE,
  ...
)
}
\arguments{
\item{vec_t}{patient enrollment time}

\item{n_interval}{number of intervals for historical enrollment}

\item{t_start}{time corresponding to the start of the study}

\item{t_hist}{time corresponding to the end of the historical study and start
of the current randomized segment}

\item{t_end}{time corresponding to the end of the study}

\item{trange}{map enrollment time to certain range}
}
\description{
Map enrollment time to [0, start of current study] and [start of current
study, end of current study]. Then, create intervals for [0, start of current
study].
}
