% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_sample.R
\name{pdSampleSingle}
\alias{pdSampleSingle}
\title{Sample with a mixture prior for binary outcomes}
\usage{
pdSampleSingle(vec_y, prior_smps = NULL, epsilon = 1,
  mix_ab = c(0.25, 0.25), iter = 5000, ...)
}
\arguments{
\item{mix_ab}{beta shape parameters (a, b) for a non-informative prior}

\item{iter}{number of thetas to draw when there is no prior samples.
otherwise, draw the same number of thetas as in the prior samples}

\item{...}{optional parameters for \code{\link{pdStan}}}

\item{vecy}{vector of binary outcomes}

\item{epsillon}{mixture proportion from the prior samples. ignored if
prior_smps is NULL}
}
\value{
samples of \eqn{theta}
}
\description{
Sample with a mixture prior for binary outcomes
}
