% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simu.R
\name{pdSimuTime}
\alias{pdSimuTime}
\title{Simulate enrollment}
\usage{
pdSimuTime(
  n_pat,
  fmla_t = NULL,
  coeff_t = 1,
  dtime = NULL,
  t_start = 0,
  t_end = 3,
  seed = NULL,
  ...
)
}
\description{
Simulate enrollment
}
