% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simu.R
\name{pdSimuX}
\alias{pdSimuX}
\title{Simulate covariates}
\usage{
pdSimuX(nPat, muCov, sdCov, corCov, cov.breaks = NULL, fmla_x = NULL,
  coeff_x = 1, dx = NULL, seed = NULL, ...)
}
\description{
Simulate covariates
}
