% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_package.R
\docType{package}
\name{pfDesign-package}
\alias{pfDesign-package}
\alias{pfDesign}
\title{The 'pfDesign' package.}
\description{
An Analysis Toolbox for Platform Design
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
