% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_simustudy.R
\name{ps_ss_get_par}
\alias{ps_ss_get_par}
\title{Get coefficients for a polynomial model to mimic a sigmoid curve}
\usage{
ps_ss_get_par(t_start, t_end, fml, L = 0.4, k = 1, t0 = 2)
}
\description{
Get coefficients for a polynomial model to mimic a sigmoid curve
}
