% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtest-package.R
\docType{package}
\name{rtest-package}
\alias{rtest-package}
\alias{rtest}
\title{The 'rtest' package.}
\description{
A template for R package that uses rstan and Rcpp
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
