% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtest_stan.R
\name{rtSTAN}
\alias{rtSTAN}
\title{Call STAN models}
\usage{
rtSTAN(lst.data, stan.mdl = "example", chains = 4, iter = 2000,
  warmup = 1000, control = list(adapt_delta = 0.95), ...)
}
\arguments{
\item{chains}{STAN parameter. Number of Markov chainsm}

\item{iter}{STAN parameter. Number of iterations}

\item{warmup}{STAN parameter. Number of burnin.}

\item{control}{STAN parameter. See \code{rstan::stan} for details.}

\item{...}{other options to call STAN sampling such as \code{thin},
\code{algorithm}. See \code{rstan::sampling} for details.#'}
}
\description{
Call STAN models
}
