% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolbox.R
\name{tkExpRst}
\alias{tkExpRst}
\title{Export results into template latex tables}
\usage{
tkExpRst(numbers, template_f, out_f, sub.str = "AA", append = FALSE)
}
\description{
Export results into template latex tables
}
