% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_sample.R
\name{pdGetInt}
\alias{pdGetInt}
\title{Create interval}
\usage{
pdGetInt(vec_t, n_interval, trange = c(0, 4))
}
\arguments{
\item{trange}{map enrollment time to certain range}
}
\description{
Create interval
}
