% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfDesign_parametric.R
\name{pdLogist}
\alias{pdLogist}
\title{Estimate response rates using parametric logistic model}
\usage{
pdLogist(dta, fml, n_bs = 500)
}
\arguments{
\item{dta}{data frame}

\item{fml}{logistic model formula}

\item{n_bs}{number of bootstraps}
}
\value{
mean and bootstrap variance of the current study response rates
}
\description{
Estimate response rates using parametric logistic model
}
